explore-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
pydantic[email]>=2.5.2
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
python-multipart>=0.0.6
alembic>=1.13.0
